:sparkles: Use concurrent map and next/done behavior change

+++ components/batch/config.go
@@ -4,6 +4,6 @@ package batcher
 type Config struct {
 	Deflate           bool
 	TimeoutMillis     uint
-	Limit             uint
+	Limit             uint64
 	MaxPendingBatches uint
 }
